idcp: Reimplement subscription management
IDCP relies on subscribing to event notification from the
IDC peers. Such subscriptions are time limited and have to
be periodically renewed.

Autobahn currently submits the renew requests to the consuming
webservice and not the broker that distributes the events.
Thus after 50 minutes (the default lifetime for subscription
requests) the subscription stops working. Since the failure
is only logged at debug log level, discovering this failure is
quite difficult.

Even if renew requests were sent to the correct web service,
the exising implementation doesn't recover from renewal failures.
Ie autobahn will keep attempting to renew a subscription when
it actually should submit a new subscription request.

Failure to renew is quite likely because Autobahn waited until
the expiration of the subscription. The spec however states the
subscription has to be renewed *before* the subscription expires.

Finally, the actual subscription lifetime is negotiated between
the two peers. Autobahn specifies a lifetime (well, currently
it doesn't, but that's a different issue) and Oscars takes this
into consideration. It then replies with the real lifetime. Renewal
should take the real lifetime into account, but the current code
doesn't.

This patch fixes these problems. It refactors subscription
management by moving the code to a new SubscriptionManager class.
This class internally takes care of subscription renewal and
resubscription in case renewal is no longer possible. It also
takes care of the situation in which the IDC peer is down while
Autobahn is started: Autobahn will now keep resubmitting the
subscription request every 30 seconds until it succeeds.

I took the liberty to also refactor parts of IdcpNotifyClient to
make the interaction witht he SubscriptionManager a bit nicer.

Testing done: Tested with Oscars 0.6 (or rather a patched version
of it). Verified by inspecting log output that reservations were
renewed. Submitted reservations after 1 hour to verify that
notifications actually worked - without this patch reservations
would fail.